changed score to a BigDecimal

+++ SquintV2/SquintV2/src/main/java/com/example/SquintV2/Models/Productivity.java
@@ -28,7 +28,7 @@ public class Productivity{
     private LocalDate date;
 
     @Column(value = "score")
-    private Integer score;
+    private BigDecimal score;
 
     public Integer getProductivityId() {
         return this.productivity_id;
